Update only on changed content.

+++ bin/post-incs.pl
@@ -20,45 +20,54 @@ foreach my $fn (@ARGV)
         return io->file($fn)->utf8;
     };
 
-    my $text = $_f->()->all;
+    eval {
+        my $orig_text = $_f->()->all;
+        my $text      = $orig_text;
 
-    # In case there's an error - fail and need to rebuild.
-    $_f->()->unlink();
-
-    if ( !$ENV{NO_I} )
-    {
-        $text =~
-            s#^\({5}include "([^"]+)"\){5}\n#io->file("lib/$1")->utf8->all#egms;
-        $text =~
-s#\({5}chomp_inc='([^']+)'\){5}#io->file("lib/$1")->chomp->utf8->getline("\n")#egms;
-    }
-
-    if ( $ENV{F} )
-    {
-        foreach my $class (qw(info irc-conversation))
+        if ( !$ENV{NO_I} )
         {
-            my $table_from = qq{<table class="$class">};
-            my $table_to   = qq{<table class="$class" summary="">};
-
-            $text =~ s#\Q$table_from\E#$table_to#g;
+            $text =~
+s#^\({5}include "([^"]+)"\){5}\n#io->file("lib/$1")->utf8->all#egms;
+            $text =~
+s#\({5}chomp_inc='([^']+)'\){5}#io->file("lib/$1")->chomp->utf8->getline("\n")#egms;
         }
-    }
 
-    # Disabling tidyp for the time being due to:
-    # https://github.com/petdance/tidyp/issues/21 .
-    if (0)
+        if ( $ENV{F} )
+        {
+            foreach my $class (qw(info irc-conversation))
+            {
+                my $table_from = qq{<table class="$class">};
+                my $table_to   = qq{<table class="$class" summary="">};
+
+                $text =~ s#\Q$table_from\E#$table_to#g;
+            }
+        }
+
+        # Disabling tidyp for the time being due to:
+        # https://github.com/petdance/tidyp/issues/21 .
+        if (0)
+        {
+            $text = $tidy->clean($text);
+        }
+
+        $text =~ s# +$##gms;
+        $text =~ s#(</div>|</li>|</html>)\n\n#$1\n#g;
+
+        # Remove surrounding space of tags.
+        $text =~
+            s#\s*(</?(?:body|(?:br /)|div|head|li|ol|p|title|ul)>)\s*#$1#gms;
+
+        # Remove document trailing space.
+        $text =~ s#\s+\z##ms;
+
+        if ( $text ne $orig_text )
+        {
+            $_f->()->print($text);
+        }
+    };
+    if ($@)
     {
-        $text = $tidy->clean($text);
+        # In case there's an error - fail and need to rebuild.
+        $_f->()->unlink();
     }
-
-    $text =~ s# +$##gms;
-    $text =~ s#(</div>|</li>|</html>)\n\n#$1\n#g;
-
-    # Remove surrounding space of tags.
-    $text =~ s#\s*(</?(?:body|(?:br /)|div|head|li|ol|p|title|ul)>)\s*#$1#gms;
-
-    # Remove document trailing space.
-    $text =~ s#\s+\z##ms;
-
-    $_f->()->print($text);
 }
